yolo-microservice/requirements.txt
fastapi
uvicorn
ultralytics
pillow
opencv-python
pytesseract
numpy